+++ app/lib/groups/src/pages/course/create/course_template.dart
