+++ app/lib/groups/src/pages/course/create/course_template_page.dart
@@ -16,7 +16,7 @@ import 'package:sharezone/groups/src/pages/course/create/my_admin_school_classes
 import 'package:sharezone/groups/src/pages/course/create/src/bloc/course_create_bloc.dart';
 import 'package:sharezone/groups/src/pages/course/create/src/bloc/course_create_bloc_factory.dart';
 import 'package:sharezone/groups/src/pages/school_class/card/school_class_card.dart';
-import 'package:sharezone/groups/src/pages/shared/course_template.dart';
+import 'package:sharezone/groups/src/pages/course/create/course_template.dart';
 import 'package:sharezone_common/api_errors.dart';
 import 'package:sharezone_widgets/sharezone_widgets.dart';
