+++ app/lib/groups/src/pages/course/create/src/analytics/course_create_analytics.dart
@@ -7,7 +7,7 @@
 // SPDX-License-Identifier: EUPL-1.2
 
 import 'package:analytics/analytics.dart';
-import 'package:sharezone/groups/src/pages/shared/course_template.dart';
+import 'package:sharezone/groups/src/pages/course/create/course_template.dart';
 
 import 'events/course_create_event.dart';
